+++ base/sparse.jl
@@ -21,7 +21,7 @@ import Base: @get!, acos, acosd, acot, acotd, acsch, asech, asin, asind, asinh,
     exp, expm1, factorize, find, findmax, findmin, findnz, float, full, getindex,
     hcat, hvcat, imag, indmax, ishermitian, kron, length, log, log1p, max, min,
     maximum, minimum, norm, one, promote_eltype, real, reinterpret, reshape, rot180,
-    rotl90, rotr90, round, scale, scale!, setindex!, similar, size, transpose, tril,
+    rotl90, rotr90, round, scale!, setindex!, similar, size, transpose, tril,
     triu, vcat, vec
 
 import Base.Broadcast: eltype_plus, broadcast_shape